+++ src/Moodle/Infrastructure/Internationalization.php
@@ -35,8 +35,13 @@ final class Internationalization
     public const VIEW_ALERT_MANY_ROOMS = 'view_alert_many_rooms';
     public const VIEW_BUTTON_JOIN_ROOM = 'view_button_join_room';
     public const VIEW_ERROR_NO_GROUPS = 'view_error_no_groups';
+    public const VIEW_WARNING_NO_MATRIX_USER_ID = 'view_error_no_matrix_user_id';
     public const VIEW_ERROR_NO_ROOM_IN_GROUP = 'view_error_no_room_in_group';
     public const VIEW_ERROR_NO_ROOMS = 'view_error_no_rooms';
     public const VIEW_ERROR_NO_VISIBLE_GROUPS = 'view_error_no_visible_groups';
     public const VIEW_HEADER = 'view_header';
+    public const VIEW_MATRIX_USER_ID_FORM_HEADER = 'view_matrix_user_id_form_header';
+    public const VIEW_MATRIX_USER_ID_FORM_NAME = 'view_matrix_user_id_form_name';
+    public const VIEW_MATRIX_USER_ID_FORM_NAME_ERROR_REQUIRED = 'view_matrix_user_id_form_name_error_required';
+    public const VIEW_MATRIX_USER_ID_FORM_NAME_HELP = 'view_matrix_user_id_form_name_help';
 }